Moved detection properties to IDetector
Redmine: #1859
