+++ Core/Instrument/IDetector.cpp
@@ -1,7 +1,16 @@
 #include "IDetector.h"
 
-IDetector::IDetector(const IDetector& other) : m_axes(other.m_axes)
+IDetector::IDetector()
 {
+    registerChild(&m_detection_properties);
+}
+
+IDetector::IDetector(const IDetector& other)
+    : m_axes(other.m_axes)
+    , m_detection_properties(other.m_detection_properties)
+{
+    setName(other.getName());
+    registerChild(&m_detection_properties);
 }
 
 void IDetector::addAxis(const IAxis& axis)
@@ -40,5 +49,31 @@ size_t IDetector::getTotalSize() const
         result *= m_axes[i_axis]->size();
     }
     return result;
+}
+
+void IDetector::setAnalyzerProperties(const kvector_t direction, double efficiency,
+                                        double total_transmission)
+{
+    m_detection_properties.setAnalyzerProperties(direction, efficiency, total_transmission);
+}
+
+kvector_t IDetector::analyzerDirection() const
+{
+    return m_detection_properties.analyzerDirection();
+}
+
+double IDetector::analyzerEfficiency() const
+{
+    return m_detection_properties.analyzerEfficiency();
+}
+
+double IDetector::analyzerTotalTransmission() const
+{
+    return m_detection_properties.analyzerTotalTransmission();
+}
+
+std::vector<const INode*> IDetector::getChildren() const
+{
+    return std::vector<const INode*>() << &m_detection_properties;
 }
